Rework

+++ src/metrics/MetricsCalculator.java
@@ -7,29 +7,29 @@ public class MetricsCalculator {
     
     public static double calculateAverageWaitingTime(List<Process> processes) {
         if (processes == null || processes.isEmpty()) return 0.0;
-        int total = 0;
+        double total = 0.0;
         for (Process p : processes) {
             total += p.getWaitingTime();
         }
-        return (double) total / processes.size();
+        return total / processes.size();
     }
     
     public static double calculateAverageTurnaroundTime(List<Process> processes) {
         if (processes == null || processes.isEmpty()) return 0.0;
-        int total = 0;
+        double total = 0.0;
         for (Process p : processes) {
             total += p.getTurnaroundTime();
         }
-        return (double) total / processes.size();
+        return total / processes.size();
     }
     
     public static double calculateAverageResponseTime(List<Process> processes) {
         if (processes == null || processes.isEmpty()) return 0.0;
-        int total = 0;
+        double total = 0.0;
         for (Process p : processes) {
             total += p.getResponseTime();
         }
-        return (double) total / processes.size();
+        return total / processes.size();
     }
     
     public static String formatComparisonTable(double rrWT, double priorityWT,
@@ -39,9 +39,9 @@ public class MetricsCalculator {
         sb.append("╔══════════════════════╦══════════════╦══════════════╗\n");
         sb.append("║ Metric               ║ Round Robin  ║ Priority     ║\n");
         sb.append("╠══════════════════════╬══════════════╬══════════════╣\n");
-        sb.append(String.format("║ Avg Waiting Time    ║ %10.2f ║ %10.2f ║\n", rrWT, priorityWT));
-        sb.append(String.format("║ Avg Turnaround Time ║ %10.2f ║ %10.2f ║\n", rrTAT, priorityTAT));
-        sb.append(String.format("║ Avg Response Time   ║ %10.2f ║ %10.2f ║\n", rrRT, priorityRT));
+        sb.append(String.format("║ Avg Waiting Time    ║ %10.4f ║ %10.4f ║\n", rrWT, priorityWT));
+        sb.append(String.format("║ Avg Turnaround Time ║ %10.4f ║ %10.4f ║\n", rrTAT, priorityTAT));
+        sb.append(String.format("║ Avg Response Time   ║ %10.4f ║ %10.4f ║\n", rrRT, priorityRT));
         sb.append("╚══════════════════════╩══════════════╩══════════════╝\n");
         return sb.toString();
     }
